requirements.txt
python_graphql_client
python-dotenv